image-captioning/requirements.txt
torch
torchvision
Pillow
nltk
